TR_query() function signature matches TR_inspect()
`observable` is not being used in this function but `text_block` is so change the function signature of `TR_query()` to match `TR_inspect()` just below.  This resolves both the unused variable name `observable` and the undefined variable name `text_block`.

The user of this example will still need to define their own `TR_enrich()` function.